% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateHistory.R
\name{animateHistory}
\alias{animateHistory}
\title{Animate karyotype history}
\usage{
animateHistory(karyoSimHistory, subsetSize = 1000, generations = NULL,
  file = NULL, interval = 1)
}
\arguments{
\item{karyoSimHistory}{A list containing the karyotype matrices at every generation}

\item{subsetSize}{The maximum number of cells per generation to display}

\item{generations}{A subset of generations to plot, rather than the entire history. The 0th generation will always be plotted.}

\item{A}{file name with either a .gif, .mp4, .swf, or .html extension to which the plot will be saved.}

\item{The}{interval between the generation in seconds.}
}
\value{
An animated ggplot.
}
\description{
This function will return a looping plot of the karyotype matrices over time.
}
\author{
Bjorn Bakker
}
