% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_cinsim.R
\name{append_cinsim}
\alias{append_cinsim}
\title{Append CINsim}
\usage{
append_cinsim(karyoSim1 = NULL, karyoSim2 = NULL)
}
\arguments{
\item{karyoSim1}{A first karyoSim object.}

\item{karyoSim2}{A second karyoSim object.}
}
\value{
A joined karyoSim object
}
\description{
This functions will enable appending two CINsim objects to be joined together. This would allow
for instance continuing simulations with a karyotype matrix from a previous simulation.
}
\author{
Bjorn Bakker
}
