% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CnFS_KMS.R
\name{calc_CnFS}
\alias{calc_CnFS}
\title{Calculate the Copy number Frequency Score (CnFS) given a set of karyotypes
and a selection metric as used by CINsim}
\usage{
calc_CnFS(karyotypes, selection_metric)
}
\arguments{
\item{karyotypes}{karyotypes as given from CINsim functions}

\item{selection_metric}{Selection metric as used in CINsim}
}
\value{
The Copy number Frequency Score (CnFS)
}
\description{
Calculate the Copy number Frequency Score (CnFS) given a set of karyotypes
and a selection metric as used by CINsim
}
\author{
Alex van Kaam
}
