% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CnFS_KMS.R
\name{calc_KMS}
\alias{calc_KMS}
\title{Calcualte KMS from karyotype & population measures
Calculate the Karyotype Matching Score (KMS) given a set of karyotypes
and observed aneuploidy/heterogeneity scores of the simulated population.}
\usage{
calc_KMS(karyotypes, pop_measures, euploid_ref = 2)
}
\arguments{
\item{karyotypes}{karyotypes as given from CINsim functions}

\item{pop_measures}{Selection metric as used in CINsim}
}
\value{
The Copy number Frequency Score (CnFS)
}
\description{
Calcualte KMS from karyotype & population measures
Calculate the Karyotype Matching Score (KMS) given a set of karyotypes
and observed aneuploidy/heterogeneity scores of the simulated population.
}
\author{
Alex van Kaam
}
