% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClonality.R
\name{calcClonality}
\alias{calcClonality}
\title{Calculate frequency of clones in a population}
\usage{
calcClonality(karyoMat, sizeInitial = NULL)
}
\arguments{
\item{karyoMat}{A karyotype matrix, with cells in rows and chromosomes in columns.}

\item{sizeInitial}{The initial size of the total cell population at the start of the simulation.}
}
\value{
A table containting the frequencies of clones.
}
\description{
This function determines the frequency of clones in a karyotype matrix, using the matrix row names as identifiers for the parental clones
}
\author{
Bjorn Bakker
}
