% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDivProb.R
\name{calcDivProb}
\alias{calcDivProb}
\title{Calculate probability of division}
\usage{
calcDivProb(karyotype, numberOfChromosomes = NULL, probDf)
}
\arguments{
\item{karyotype}{A single karyotype as a vector.}

\item{numberOfChromosomes}{The number of chromosomes in the vector.}

\item{probDf}{A probability fitness matrix.}
}
\value{
A probability of cell division.
}
\description{
This function will return a probability of division, based on a probability fitness matrix, such that cells with higher fitness have a greater chance of dividing. Whenever the calculated probability is less than 0.7, a value of 0 is returned instead.
}
\author{
Bjorn Bakker
}
