% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMisseg.R
\name{calcMisseg}
\alias{calcMisseg}
\title{Calculate adjusted probability of mis-segregation}
\usage{
calcMisseg(karyotype, pMisseg, pMissegCoef, numberOfChromosomes)
}
\arguments{
\item{karyotype}{A single karyotype, supplied as a vector.}

\item{pMisseg}{The probability of mis-segregation per chromosome copy.}

\item{pMissegCoef}{The probability co-efficient, calculated from the euploid reference karyotype.}

\item{numberOfChromosomes}{The number of chromosomes in the provided karyotype.}
}
\value{
A probability of chromosome mis-segregation.
}
\description{
This function calculates the probability of mis-segregation, such that cells with higher fitness have a lower probability of mis-segregating.
}
\author{
Bjorn Bakker
}
