% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurProb.R
\name{calcSurProb}
\alias{calcSurProb}
\title{Calculate probability of survival in a fitness-dependent manner}
\usage{
calcSurProb(karyotype, numberOfChromosomes = NULL, probDf = NULL)
}
\arguments{
\item{karyotype}{A karyotype provided as a vector.}

\item{numberOfChromosomes}{The number of chromosomes in the karyotype.}

\item{probDf}{A fitness probability matrix.}
}
\value{
A probability of survival.
}
\description{
This function will calculate a probability of survival based on karyotypic fitness.
}
\author{
Bjorn Bakker
}
