% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld_function.R
\name{calculate_kld}
\alias{calculate_kld}
\title{Determine the karyotype landscape dissimilarity (KLD) score}
\usage{
calculate_kld(karyoSim, km_reference = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim object containing simulation information.}

\item{km_reference}{A table that contains the karyotype measures (aneuploidy and heterogeneity score) per chromosome.}
}
\value{
A data frame with the karyotype dissimilarity index (KLD) per generation (absolute value and normalized to g0).
}
\description{
This function returns a score that quantifies the  dissimilarity of a karyotype landscape compared to observed
aneuploidy and heterogeneity scores (on a chromosome basis). It can be used to determine whether results of a
simulation yield a karyotype landscape that resembles a tumour sample.
}
\author{
Bjorn Bakker
}
