% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_viability.R
\name{check_viability}
\alias{check_viability}
\title{Check viability of a simulation
Determine whether a simulation is viable by checking whether
the maximum amount of cells or generations was reached.}
\usage{
check_viability(sim_list, sim_number = 1:100, threshold_value, max_g)
}
\arguments{
\item{sim_list}{list of simulations as gotten from parallelCinsim}

\item{sim_number}{indication of how many simulations are in the sim_list}

\item{threshold_value}{the amount cells before we consider a simulation viable}

\item{max_g}{The amount of generations before we consider a simulation viable}

\item{pop_measures}{Selection metric as used in CINsim}
}
\value{
TRUE/FALSE whether a simulation is viable
}
\description{
Check viability of a simulation
Determine whether a simulation is viable by checking whether
the maximum amount of cells or generations was reached.
}
\author{
Alex van Kaam
}
