% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnFreq.R
\name{cnFreq}
\alias{cnFreq}
\title{Calculate copy number frequency}
\usage{
cnFreq(karyoSim, plot = TRUE)
}
\arguments{
\item{karyoSim}{A karyoSim object (final output of the CINsim function).}

\item{plot}{A logical, whether to plot the result or not.}
}
\value{
A table or plot with the frequencies of the chromosome copy number states.
}
\description{
This function will determine the frequencies of copy numbers per chromosomes, and return either a table or a barplot.
}
\author{
Bjorn Bakker
}
