% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_iterative_karyoSimParallel}
\alias{compile_iterative_karyoSimParallel}
\title{Compile the metrics of multiple karyoSimParallel objects within distinct inputs (i.e. non-replicates)}
\usage{
compile_iterative_karyoSimParallel(karyoSimParallel, sim_label = NULL)
}
\arguments{
\item{karyoSimParallel}{A list of karyoSimParallel objects.}

\item{sim_label}{A custom label for the sets of simulations.}
}
\value{
A list of four tibbles.
}
\description{
This functions will enable easy transformation of parallel Cinsim data into tidy compiled data
that is compatible with the tidyverse functions and plotting in ggplot2. This function is intended for
parallel simulations that each have a unique parameter (e.g. they have looped over values for pMisseg).
}
\author{
Bjorn Bakker
}
