% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_karyoSim}
\alias{compile_karyoSim}
\title{Compile data from CINsim}
\usage{
compile_karyoSim(karyoSim, sim_label = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim object.}

\item{sim_label}{A custom label for the simulation.}
}
\value{
A list of three tibbles.
}
\description{
This functions will enable easy transformation of parallel Cinsim data into tidy data
that is compatible with the tidyverse functions and plotting in ggplot2.
}
\author{
Bjorn Bakker
}
