% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cinsim_simple.R
\name{compile_karyoSim_simple}
\alias{compile_karyoSim_simple}
\title{Compile data from a list of parallel simulations (simplified)}
\usage{
compile_karyoSim_simple(karyoSim, sim_label = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim_simple or karyoSimParallel_simple object.}

\item{sim_label}{A custom label for the sets of simulations.}
}
\value{
A list of three tibbles.
}
\description{
This functions will enable easy transformation of parallel Cinsim data into tidy data
that is compatible with the tidyverse function and plotting in ggplot2.
}
\author{
Bjorn Bakker
}
