% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_karyoSimParallel}
\alias{compile_karyoSimParallel}
\title{Compile data from a list of parallel simulations}
\usage{
compile_karyoSimParallel(
  karyoSimParallel,
  sim_label = NULL,
  for_summary = FALSE
)
}
\arguments{
\item{karyoSimParallel}{A karyoSimParallel object.}

\item{sim_label}{A custom label for the sets of simulations.}

\item{for_summary}{A logical whether the data will be summarized later.}
}
\value{
A list of three tibbles.
}
\description{
This functions will enable easy transformation of parallel Cinsim data into tidy data
that is compatible with the tidyverse functions and plotting in ggplot2.
}
\author{
Bjorn Bakker
}
