% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score2prob.R
\name{create_modal_karyotype}
\alias{create_modal_karyotype}
\title{Creates a cell with modal karyotype}
\usage{
create_modal_karyotype(selection_metric)
}
\arguments{
\item{selection_metric}{The copy number frequency matrix.}
}
\value{
A single cell with modal copy number for each chromosome
}
\description{
This function creates a cell with modal copy number, given a selection metric as
used within the CINsim package
}
\author{
Alex van Kaam
}
