% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype_measures.R
\name{dev_from_mode}
\alias{dev_from_mode}
\title{Calculate the fraction of cells that deviate from the modal copy number}
\usage{
dev_from_mode(karyotypes)
}
\arguments{
\item{karyotypes}{A karyotype matrix with cells in rows and chromosomes in columns.}
}
\value{
A vector of deviation from modal scores per chromosome.
}
\description{
This function will determine the fraction of cells that deviate from the model copy number of that chromosome.
}
\author{
Bjorn Bakker
}
