% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devFromMode.R
\name{devFromMode}
\alias{devFromMode}
\title{Calculate the fraction of cells that deviate from the modal copy number}
\usage{
devFromMode(karyoSim, plot = FALSE, plotTitle = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim object.}

\item{plot}{Whether to plot the results or return a matrix.}

\item{plotTitle}{A title for the plot.}
}
\value{
A named vector or plot showing the deviation from modal chromosome copy number.
}
\description{
This function will determine the fraction of cells that deviate from the model copy number of that chromosome.
}
\author{
Bjorn Bakker
}
