% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMitosis.R
\name{doMitosis}
\alias{doMitosis}
\title{Perform a round of mitosis with possible mis-segregations}
\usage{
doMitosis(karyotype, pMisseg, cellID)
}
\arguments{
\item{karyotype}{A karyotype provided as a vector on which to apply the mitosis.}

\item{pMisseg}{The probability of mis-segregation per chromosome copy number.}

\item{cellID}{The cell ID of the provided karyotype to ensure faithful clone tracking.}
}
\value{
A 2-row matrix with karyotypes from daughter cells.
}
\description{
This function will perform mitosis on a given karyotype, apply chromosome mis-segregation events if the criteria are met, and returns a matrix with two new karytoypes.
}
\author{
Bjorn Bakker
}
