% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CnFS_KMS.R
\name{get_final_CnFS}
\alias{get_final_CnFS}
\title{Calculate the final Copy number Frequency Score for a karyoSim object.}
\usage{
get_final_CnFS(sim_results)
}
\arguments{
\item{sim_results}{A karyoSim object from CINsim functions}
}
\value{
The Copy number Frequency Score (CnFS) for the final
generation for the given simulation
}
\description{
Calculate the final Copy number Frequency Score for a karyoSim object.
}
\author{
Bjorn Bakker, Alex van Kaam
}
