% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score2prob.R
\name{get_score}
\alias{get_score}
\title{Convert a karyotype score into a probability of survival}
\usage{
get_score(
  karyotypes = NULL,
  selection_mode = NULL,
  selection_metric = NULL,
  chrom_weights = NULL,
  monosomy_penalty = FALSE,
  penalty_fraction = 0.1
)
}
\arguments{
\item{selection_mode}{One of three possible modes, "cn_based", "rel_copy", "davoli", for copy number-based, relative copy numbed-based or Davoli-based selection.}

\item{selection_metric}{A fitness metric for the cn_based, rel_copy or davoli selection measure.}

\item{chrom_weights}{An optional vector of chromosome weights to affect cellular fitness.}

\item{monosomy_penalty}{Boolean whether to apply extra score penalty for rare/non-modal monosomies.}

\item{penalty_fraction}{Expressed as a fraction between 0 and 1, how much of the maximum possible score a non-modal monosomy takes as penalty.}

\item{karyotyepes}{A matrix of karyotypes with chromosomes in columns and cells in rows.}
}
\value{
A fitness score.
}
\description{
This functions calculate the survival probability given a karyotype score according to specified coefficients.
}
\author{
Bjorn Bakker, Alex van Kaam
}
