% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getMode}
\alias{getMode}
\title{Determine the mode in a vector of values.}
\usage{
getMode(x)
}
\arguments{
\item{x}{The vector from which to extract the mode.}
}
\value{
The mode of a vector.
}
\description{
This function returns the mode (most frequent element) of a vector. R does not have an in-built function to do this.
}
\author{
Bjorn Bakker
}
