% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{karyotype_survival}
\alias{karyotype_survival}
\title{Determine cell survival}
\usage{
karyotype_survival(
  karyotypes,
  copy_num_boundaries = c(1, 8),
  selection_mode = NULL,
  selection_metric = NULL,
  a = NULL,
  b = NULL,
  chrom_weights = NULL,
  max_monosomy = NULL,
  min_euploid = NULL,
  euploid_ref = 2,
  monosomy_penalty = FALSE,
  penalty_fraction = 0.1,
  qMod = 1
)
}
\arguments{
\item{karyotypes}{A matrix with cells in rows and chromosomes in colums. Rows and colums must be labelled appropriately.}

\item{copy_num_boundaries}{A vector of length 2 with the minimum and maximum viable copy number state for any chromosome set.}

\item{selection_mode}{One of three possible modes, "cn_based", "rel_copy", "davoli", for copy number-based, relative copy numbed-based or Davoli-based selection.}

\item{selection_metric}{A fitness metric for the cn_based, rel_copy or davoli selection measure.}

\item{a}{Value of slope for karyotype selection.}

\item{b}{Value of intercept for karyotype selection.}

\item{chrom_weights}{A vector with chromosome weights to scale importance of one or more chromosome sets.}

\item{max_monosomy}{The maximum number of allowed chromosome sets to be monosomic.}

\item{min_euploid}{The minimum number of required chromosome sets to be euploid.}

\item{euploid_ref}{The euploid copy number state.}

\item{monosomy_penalty}{Boolean whether to apply extra score penalty for non-modal monosomies.}

\item{penalty_fraction}{Expressed as a fraction between 0 and 1, how much of the maximum possible score a non-modal monosomy takes as penalty.}

\item{qMod}{A modifier for the survival probability.}
}
\value{
A vector of logicals whether the cells have survived or not.
}
\description{
This function will determine the survival probability for a karyotype, depending on specific parameters.
}
\author{
Bjorn Bakker, Alex van Kaam
}
