% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKaryotypes.R
\name{makeKaryotypes}
\alias{makeKaryotypes}
\title{Make a matrix of karyotypes}
\usage{
makeKaryotypes(numCell = 100, species = "mouse", copies = 2)
}
\arguments{
\item{numCell}{The desired number of cells in the matrix.}

\item{species}{The desired species, either mouse or human (20 or 23 chromosomes).}

\item{copies}{The number of copies for all chromosomes in the karyotype matrix.}
}
\value{
A matrix of karyotypes according to the input parameters.
}
\description{
This function will create a matrix of karyotypes that is compatible with the CINsim pipeline. Currently supported species are mouse (20 chromosomes) and human (23 chromsomes).
}
\author{
Bjorn Bakker
}
