% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cinsim_simple.R
\name{parallelCinsim_simple}
\alias{parallelCinsim_simple}
\title{Run multiple simulations in parallel}
\usage{
parallelCinsim_simple(
  iterations = 6,
  num_free_cpu = 1,
  karyotypes = NULL,
  g = 12,
  pMisseg = 0.0025,
  pWGD = 0,
  copy_num_boundaries = c(1, 8),
  selection_mode = NULL,
  selection_metric = NULL,
  coef = NULL,
  qMod = NULL,
  down_sample = 50000,
  down_sample_frac = 0.25,
  max_num_cells = 2e+09,
  collect_fitness_score = FALSE
)
}
\arguments{
\item{iterations}{The number of replicate iterations to generate.}

\item{num_free_cpu}{The number of CPUs to keep available during processing.}

\item{karyotypes}{A single matrix or list of matrices with cells in rows and chromosomes in colums. Rows and colums must be labelled appropriately.}

\item{g}{The maximum number of generations to be simulated.}

\item{pMisseg}{The mis-segregation probability per chromosome copy.}

\item{pWGD}{The probability of a cell undergoing whole-genome duplication instead of a normal mitosis}

\item{copy_num_boundaries}{A minimum and maximum number of copy numbers allowed per chromosome before a cell is deemed unviable.}

\item{selection_mode}{One of three possible modes, "cn_based", "rel_copy", "davoli", for copy number-based, relative copy numbed-based or Davoli-based selection.}

\item{selection_metric}{A fitness metric for the cn_based, rel_copy or davoli selection measure.}

\item{coef}{A vector of length three containing the selection coefficients (slope and intercept and constant).}

\item{qMod}{A value modify the calculated value for pSurvival.}

\item{down_sample}{The maximum size of the simulated population before down-sampling.}

\item{down_sample_frac}{The fraction of cells with which to continue the simulation after down-sampling.}

\item{max_num_cells}{The maximum number of (theoretical) cells to simulate before the simulation is terminated.}

\item{collect_fitness_score}{A logical whether to collect the fitness scores over time.}
}
\value{
A \code{list} with karyoSim objects.
}
\description{
This function will enable parallel simulations to be run to quickly generate replicates for particular conditions.
}
\author{
Bjorn Bakker
}
