% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_measure_plots.R
\name{plot_cinsim_summary}
\alias{plot_cinsim_summary}
\title{Plot population measures}
\usage{
plot_cinsim_summary(
  karyoSim,
  sim_label = NULL,
  summarize = TRUE,
  per_chromosome = TRUE,
  plot = FALSE
)
}
\arguments{
\item{karyoSim}{A karyoSim, karyoSimParallel, or list of karyoSimParallel objects.}

\item{sim_label}{A label for karyoSimParallel data.}

\item{summarize}{A logical whether to summarize a karyoSimParallel object - karyoSimParallel lists will always be summarized.}

\item{per_chromosome}{A logical whether to plot data per chromosome or genome-wide.}

\item{plot}{A logical whether to print the plot, or return a list of plots.}
}
\value{
A summary statistics plot.
}
\description{
This functions will plot the various population measures over time.
}
\author{
Bjorn Bakker
}
