% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clonality.R
\name{plot_clonality}
\alias{plot_clonality}
\title{Plot clonal frequencies (tidy)}
\usage{
plot_clonality(karyoSim, line_graph = TRUE)
}
\arguments{
\item{karyoSim}{A karyoSim or karyoSimParallel object that contains a summary matrix for clonality frequency per generations.}

\item{line_graph}{A logical whether to plot frequencies as a line graph or a bar graph.}
}
\value{
A table or plot of the clonal frequencies.
}
\description{
This function will allow you to plot the frequencies of the clones over time.
}
\author{
Bjorn Bakker
}
