% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cn.R
\name{plot_cn}
\alias{plot_cn}
\title{Plot CN frequencies over time}
\usage{
plot_cn(karyoSim, final_g = TRUE)
}
\arguments{
\item{karyoSim}{A karyoSim or karyoSimParallel object.}

\item{final_g}{A logical whether to plot the CN frequencies at the final generation.}
}
\value{
A plot of copy number frequencies.
}
\description{
This functions will plot the frequencie of copy numbers over time.
}
\author{
Bjorn Bakker
}
