% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cinsim_simple.R
\name{plot_cn_simple}
\alias{plot_cn_simple}
\title{Plot CN frequencies (simplified)}
\usage{
plot_cn_simple(karyoSim)
}
\arguments{
\item{karyoSim_simple}{A karyoSim_simple or karyoSimParallel_simple object.}
}
\value{
A plot of copy number frequencies.
}
\description{
This functions will plot the frequencie of copy numbers.
}
\author{
Bjorn Bakker
}
