% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_misseg_freq.R
\name{plot_misseg_freq}
\alias{plot_misseg_freq}
\title{Plot mis-segregation frequencies (tidy)}
\usage{
plot_misseg_freq(karyoSim, plot = TRUE)
}
\arguments{
\item{karyoSim}{A karyoSim object that contains information of mis-segregation frequencies.}

\item{plot}{A logical wether to plot the results or return a list of data frame.}
}
\value{
A plot of the mis-segregation frequencies
}
\description{
This function will plot the mis-segregation frequencies over time.
}
\author{
Bjorn Bakker
}
