% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClonality.R
\name{plotClonality}
\alias{plotClonality}
\title{Plot clonal frequencies}
\usage{
plotClonality(karyoSim, plot = TRUE, file = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim object that contains a summary matrix for clonality frequency per generations.}

\item{plot}{A logical whether to plot the results or return a table.}

\item{file}{File name (without extension) to which to print the plot.}
}
\value{
A table or plot of the clonal frequencies.
}
\description{
This function will allow you to plot the frequencies of the clones over time.
}
\author{
Bjorn Bakker
}
