% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDevFromMode.R
\name{plotDevFromMode}
\alias{plotDevFromMode}
\title{Plot deviation from mode per chromosome.}
\usage{
plotDevFromMode(karyoSim)
}
\arguments{
\item{karayoSim}{A karyoSim object that contains a karyotype matrix.}
}
\value{
A plot of chromosomal deviation from mode.
}
\description{
This function will plot the frequency at which cells deviate from the modal copy number per chromosome.
}
\author{
Bjorn Bakker
}
