% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstMisseg.R
\name{plotFirstMisseg}
\alias{plotFirstMisseg}
\title{Plot degree of contribution to deviation from first mis-segregation}
\usage{
plotFirstMisseg(karyoSimParallel, plot = TRUE)
}
\arguments{
\item{karyoSimParallel}{A list of karyoSim objects that contains a summary matrix for clonality frequency per generations.}

\item{plot}{A logical whether to plot the results or return a table.}
}
\value{
A plot or a tibble.
}
\description{
This function will plot the average (-/+ SD) deviation from the modal copy number as a function of the
generation of first mis-segregation. For easy interpretation of the results, all provided karyoSim
objects should be replicates of each other.
}
\author{
Bjorn Bakker
}
