% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_measures.R
\name{population_measures}
\alias{population_measures}
\title{Gather population measures}
\usage{
population_measures(karyotypes = NULL, euploid_ref = 2, g = NULL)
}
\arguments{
\item{karyotypes}{A matrix with cells in rows and chromosomes in colums. Rows and colums must be labelled appropriately.}

\item{euploid_ref}{The euploid copy number (user for aneuploidy score calculations - 2 by default).}

\item{g}{The generation identifier to add to the data frame.}
}
\value{
A data frame containing all relevant population measures of the simulation.
}
\description{
This function will collect scores for aneuploidy and heterogeneity, and determine the
fraction of cells deviating from the modal copy number.
}
\author{
Bjorn Bakker
}
