% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qAneuploidy.R
\name{qAneuploidy}
\alias{qAneuploidy}
\title{Quantify levels of aneuploidy}
\usage{
qAneuploidy(karyoMat, numberOfCells = NULL, euploidRef = NULL)
}
\arguments{
\item{karyoMat}{A karyotype matrix with cells in rows and chromosomes in columns.}

\item{numberOfCells}{The number of cells in the karyotype matrix.}

\item{euploidRef}{The reference euploid copy number. Defaults to 2 if none is provided.}
}
\value{
A vector of aneuploidy scores per chromosome.
}
\description{
This function will quantify the degree of aneuploidy in a sample (calculated as the absolute deviation from euploid). Higher scores indicate more aneuploidy.
}
\author{
Bjorn Bakker
}
