% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qHeterogeneity.R
\name{qHeterogeneity}
\alias{qHeterogeneity}
\title{Quantify levels of heterogeneity}
\usage{
qHeterogeneity(karyoMat)
}
\arguments{
\item{karyoMat}{A karyotype matrix with cells in rows and chromosomes in columns.}
}
\value{
A vector of heterogeneity scores per chromosome.
}
\description{
This function will caculate the degree of heterogeneity in a sample. Higher scores indicate greater heterogeneity.
}
\author{
Bjorn Bakker
}
