% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qHeterogeneityPar.R
\name{qHeterogeneityPar}
\alias{qHeterogeneityPar}
\title{Quantify levels of heterogeneity (parallel variant)}
\usage{
qHeterogeneityPar(karyoMat, numberOfCells = NULL, cl)
}
\arguments{
\item{karyoMat}{A karyotype matrix with cells in rows and chromosomes in columns.}

\item{numberOfCells}{The number of cells in the karyotype matrix.}

\item{cl}{A defined cluster that will be used to compute heterogeneity in a parallel fashion.}
}
\value{
A vector of heterogeneity scores per chromosome.
}
\description{
This function will caculate the degree of heterogeneity in a sample. Higher scores indicate greater heterogeneity. This is the same function as qHeterogeneity, but uses parallel computing to speed up calculations.
}
\author{
Bjorn Bakker
}
