% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtnorm.R
\name{rtnorm}
\alias{rtnorm}
\title{Restricted 'normal' distribution}
\usage{
rtnorm(n, mean = 2, sd = 1, a = 1, b = 8)
}
\arguments{
\item{n}{The number of values to generate.}

\item{mean}{The mean (defaults to 2).}

\item{sd}{The standard deviation (defaults to 1)}

\item{a}{The lower limit (default is 1)}

\item{b}{The upper limit (default is 8)}
}
\description{
This function will generate integer values based on a restricted 'normal' distribution with a lower and upper boundary (1 to 8)
}
