% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_karyotype_landscape.R
\name{scan_karyotype_landscape}
\alias{scan_karyotype_landscape}
\title{Automated scanning of karyotype landscape}
\usage{
scan_karyotype_landscape(
  experiment_name = "karyotype_landscape_scan",
  iterations = 6,
  num_free_cpu = 1,
  karyotypes = NULL,
  euploid_ref = 2,
  g = 50,
  pMisseg = 10^(seq(-6, -1, length.out = 30)),
  pWGD = 0,
  pMissegG = NULL,
  fit_misseg = FALSE,
  pDivision = 1,
  fit_division = FALSE,
  copy_num_boundaries = c(1, 8),
  selection_mode = NULL,
  selection_metric = NULL,
  min_survival_euploid = 0.1,
  max_survival_euploid = 0.9,
  max_survival = 1,
  interval = 0.1,
  probability_types = c("pSurvival"),
  chrom_weights = NULL,
  qMods = NULL,
  max_monosomy = NULL,
  min_euploid = NULL,
  down_sample = 50000,
  down_sample_frac = 0.25,
  max_num_cells = 1e+10
)
}
\arguments{
\item{experiment_name}{The name of the experiment and the folder name in which to save the results}

\item{iterations}{The number of replicates to run per condition.}

\item{karyotypes}{A matrix with cells in rows and chromosomes in colums. Rows and colums must be labelled appropriately.}

\item{euploid_ref}{The euploid chromosome copy number (used for calculating levels of aneuploidy).}

\item{g}{The maximum number of generations to be simulated.}

\item{pMisseg}{The range mis-segregation probabilities per chromosome copy.}

\item{pWGD}{The probability of a cell undergoing whole-genome duplication instead of a normal mitosis}

\item{pMissegG}{A vector of generation numbers during which pMisseg is set at the desired values. Will otherwise default to 0.}

\item{fit_misseg}{A logical whether pMisseg should be fitness-dependent.}

\item{pDivision}{The base probability of cell division (acts as a threshold for permanent cell arrest when division is fitness dependent).}

\item{fit_division}{A logical whether pDivision should be karyotype (i.e. fitness) dependent.}

\item{copy_num_boundaries}{A minimum and maximum number of copy numbers allowed per chromosome before a cell is deemed unviable.}

\item{selection_mode}{One of three possible modes, "cn_based", "rel_copy", "davoli", for copy number-based, relative copy numbed-based or Davoli-based selection.}

\item{selection_metric}{A fitness metric for the cn_based, rel_copy or davoli selection measure.}

\item{min_survival_euploid}{The minimum probability of survival for a euploid karyotype.}

\item{max_survival_euploid}{The maximum probability of survival for a euploid karyotype.}

\item{max_survival}{Maximum survival probability for the optimal karyotype score.}

\item{interval}{The interval of survival probabilities (range min to max).}

\item{probability_types}{The probabilities over which to iterate (either one of pDivision, pMisseg, pSurvival).}

\item{chrom_weights}{An optional vector of chromosome weights to affect cellular fitness.}

\item{qMods}{A vector of length 3 to modify the calculated values for pDivison, pMisseg and pSurvival respectively.}

\item{max_monosomy}{The minimum number of monosomies that will cause the cell to die.}

\item{min_euploid}{The minimum number of euploid chromosomes that must remain before the cell dies.}

\item{down_sample}{The maximum size of the simulated population before down-sampling.}

\item{down_sample_frac}{The fraction of cells with which to continue the simulation after down-sampling.}

\item{max_num_cells}{The maximum number of (theoretical) cells to simulate before the simulation is terminated.}
}
\value{
A list of relevant simulation data. KaryoSim objects are saved to file.
}
\description{
This wrapper function will allow for simulations across a range of pMisseg and selection strengths.
}
\author{
Bjorn Bakker
}
