% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score2prob.R
\name{score2prob}
\alias{score2prob}
\title{Convert a karyotype score into a probability of survival}
\usage{
score2prob(score, a, b, linear = TRUE)
}
\arguments{
\item{score}{A karyotype score.}

\item{a}{A slope coefficient.}

\item{b}{An intercept coefficient.}
}
\value{
A survival probability.
}
\description{
This functions calculate the survival probability given a karyotype score according to specified coefficients.
}
\author{
Bjorn Bakker
}
