% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPlots.R
\name{simPlots}
\alias{simPlots}
\title{Simulation metrics plot}
\usage{
simPlots(karyoSim, file = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim object that contains all the required information and metrics.}

\item{file}{A file name (without extension) to which the data shall be printed.}
}
\description{
This function will plot various simulation metrics into a single overview. Metrics include the cummulative cell number, fraction of cells surviving per round of cell division, karyotype measures (aneuploidy and heterogeneity), and clonal frequencies when the simulation starting population was greater than 1.
}
\author{
Bjorn Bakker
}
