% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedMessage.R
\name{startTimedMessage}
\alias{startTimedMessage}
\title{Start timed message}
\usage{
startTimedMessage(...)
}
\arguments{
\item{...}{Text for message}
}
\description{
Sets up a timed message to monitor simulation progression.
}
