% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedMessage.R
\name{stopTimedMessage}
\alias{stopTimedMessage}
\title{Stop a timed message}
\usage{
stopTimedMessage(ptm)
}
\arguments{
\item{ptm}{A started time message}
}
\description{
Sets up a timed message to monitor simulation progression.
}
