% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{summarize_karyoSimParallel}
\alias{summarize_karyoSimParallel}
\title{Summarize population metrics of a karyoSimParallel object}
\usage{
summarize_karyoSimParallel(karyoSimParallel, sim_label = NULL)
}
\arguments{
\item{karyoSimParallel}{A karyoSimParallel objects.}

\item{sim_label}{A custom label for the sets of simulations.}
}
\value{
A list of three tibbles.
}
\description{
This functions will enable easy transformation of parallel Cinsim data into tidy summarized data
that is compatible with the tidyverse functions and plotting in ggplot2.
}
\author{
Bjorn Bakker
}
