% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_karyotype_landscape.R
\name{summarize_karyotype_landscape_scan}
\alias{summarize_karyotype_landscape_scan}
\title{Summarizing karyotype landscape scanning output}
\usage{
summarize_karyotype_landscape_scan(karyotype_landscape_scan)
}
\arguments{
\item{karyotype_landscape_scan}{An object of class karyotype_landscape_scan.}
}
\value{
Plots and objects saved to file
}
\description{
This function will generate a report of the karyotype landscape scanning.
}
\author{
Bjorn Bakker
}
