% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CnFS.R
\name{calc_KMS}
\alias{calc_KMS}
\title{Calculate the Karyotype Matching Score (KMS) given a set of karyotypes
and a selection metric as used by CINsim}
\usage{
calc_KMS(karyotypes, pop_measures)
}
\arguments{
\item{karyotypes}{karyotypes as given from CINsim functions}

\item{selection_metric}{Selection metric as used in CINsim}
}
\value{
The Copy number Frequency Score (CnFS)
}
\description{
Calculate the Karyotype Matching Score (KMS) given a set of karyotypes
and a selection metric as used by CINsim
}
\author{
Alex van Kaam
}
