% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINsim.R
\name{Cinsim}
\alias{Cinsim}
\title{Simulate chromosomal instability.}
\usage{
Cinsim(karyotypes = NULL, euploidRef = NULL, g = 12, pMisseg = 0.01,
  pMissegG = NULL, fitMisseg = FALSE, pDivision = 0.8,
  fitDivision = FALSE, probDf = NULL, karyotypeSelection = FALSE,
  minMonosomyLethal = 3, minNumEuploidChr = 0, downSample = 25000,
  maxNumCells = 2e+09, simTitle = "CINsim", saveToFile = FALSE,
  numFreeCPU = 1)
}
\arguments{
\item{karyotypes}{A matrix with cells in rows and chromosomes in colums. Rows and colums must be labelled appropriately.}

\item{euploidRef}{The euploid chromosome copy number (used for calculating levels of aneuploidy).}

\item{g}{The maximum number of generations to be simulated.}

\item{pMisseg}{The mis-segregation probability per chromosome copy.}

\item{pMissegG}{A vector of generation numbers during which pMisseg is set at the desired values. Will otherwise default to 0.}

\item{fitMisseg}{A logical whether pMisseg should be fitness-dependent.}

\item{pDivision}{The base probability of cell division.}

\item{fitDivision}{A logical whether pDivision should be karyotype (i.e. fitness) dependent.}

\item{probDf}{A probability fitness matrix (max 8 rows for the max 8 copy number states).}

\item{karyotypeSelection}{Whether karyotype selection should take place.}

\item{minMonosomyLethal}{The minimum number of monosomies that will cause the cell to die.}

\item{minNumEuploidChr}{The minimum number of euploid chromosomes that must remain before the cell dies.}

\item{downSample}{The maximum size of the simulated population before down-sampling.}

\item{maxNumCells}{The maximum number of (theoretical) cells to simulate before the simulation is terminated.}

\item{simTitle}{A custom title linked to the final output when saving to file.}

\item{saveToFile}{A logical whether to the final simulation ouput to file.}

\item{numFreeCPU}{The number of CPUs to keep free for use when the simulation is running.}
}
\value{
A karyoSim object containing all relevant information of the simulation.
}
\description{
This the the main function for the CINsim package, that will allow for simulation of cell divisions and chromosomal instability.
}
\author{
Bjorn Bakker
}
