% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cinsim.R
\name{Cinsim}
\alias{Cinsim}
\title{Simulate chromosomal instability.}
\usage{
Cinsim(
  karyotypes = NULL,
  euploid_ref = 2,
  g = 12,
  pMisseg = 0.0025,
  pWGD = 0,
  pMissegG = NULL,
  fit_misseg = FALSE,
  pDivision = 1,
  fit_division = FALSE,
  copy_num_boundaries = c(1, 8),
  selection_mode = NULL,
  selection_metric = NULL,
  coef = NULL,
  chrom_weights = NULL,
  qMods = NULL,
  max_monosomy = NULL,
  min_euploid = NULL,
  down_sample = 50000,
  down_sample_frac = 0.25,
  max_num_cells = 2e+09,
  collect_fitness_score = FALSE,
  CnFS = TRUE,
  KMS = FALSE,
  final_aneu_het_scores = NULL,
  verbose = 1
)
}
\arguments{
\item{karyotypes}{A matrix with cells in rows and chromosomes in colums. Rows and colums must be labelled appropriately.}

\item{euploid_ref}{The euploid chromosome copy number (used for calculating levels of aneuploidy).}

\item{g}{The maximum number of generations to be simulated.}

\item{pMisseg}{The mis-segregation probability per chromosome copy.}

\item{pWGD}{The probability of a cell undergoing whole-genome duplication instead of a normal mitosis}

\item{pMissegG}{A vector of generation numbers during which pMisseg is set at the desired values. Will otherwise default to 0.}

\item{fit_misseg}{A logical whether pMisseg should be fitness-dependent.}

\item{pDivision}{The base probability of cell division (acts as a threshold for permanent cell arrest when division is fitness dependent).}

\item{fit_division}{A logical whether pDivision should be karyotype (i.e. fitness) dependent.}

\item{copy_num_boundaries}{A minimum and maximum number of copy numbers allowed per chromosome before a cell is deemed unviable.}

\item{selection_mode}{One of three possible modes, "cn_based", "rel_copy", "davoli", for copy number-based, relative copy numbed-based or Davoli-based selection.}

\item{selection_metric}{A fitness metric for the cn_based, rel_copy or davoli selection measure.}

\item{coef}{A list of length three containing the karyotype fitness coefficients (slope and intercept).}

\item{chrom_weights}{An optional vector of chromosome weights to affect cellular fitness.}

\item{qMods}{A vector of length 3 to modify the calculated values for pDivison, pMisseg and pSurvival respectively.}

\item{max_monosomy}{The minimum number of monosomies that will cause the cell to die.}

\item{min_euploid}{The minimum number of euploid chromosomes that must remain before the cell dies.}

\item{down_sample}{The maximum size of the simulated population before down-sampling.}

\item{down_sample_frac}{The fraction of cells with which to continue the simulation after down-sampling.}

\item{max_num_cells}{The maximum number of (theoretical) cells to simulate before the simulation is terminated.}

\item{collect_fitness_score}{A logical whether to collect the fitness scores over time.}

\item{CnFS}{Boolean to indicate whether the CnFS should be calculated, defaults to TRUE}

\item{KMS}{Boolean to indicate whether the KMS should be calculated, defaults to FALSE}

\item{final_aneu_het_scores}{The final population measures (heterogeneity & aneuploidy) required for calculating the KMS.}

\item{verbose}{Degree of verbosity of CINsim, 0 means no verbosity}
}
\value{
A karyoSim object containing all relevant information of the simulation.
}
\description{
This the the main function for the CINsim package, that will allow for simulation of cell divisions and chromosomal instability.
}
\author{
Bjorn Bakker
}
