% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvHeatmap.R
\name{cnvHeatmap}
\alias{cnvHeatmap}
\title{Plot a copy number heatmap.}
\usage{
cnvHeatmap(karyoSim = NULL, subset_size = 1000, clones = NULL, file = NULL)
}
\arguments{
\item{karyoSim}{A karyoSim object (final output of CINsim).}

\item{subset_size}{The number of cells to subset and plot from the karyotype matrix.}

\item{clones}{A character vector of the clone identifiers to plot.}

\item{file}{A (optional) name for the pdf-file to which the plot will be printed.}
}
\value{
A heatmap with copy numbers shown in colours.
}
\description{
This function will plot a copy number heatmap based on a karyotype matrix.
}
\author{
Bjorn Bakker
}
