% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKaryotypes.R
\name{makeKaryotypes}
\alias{makeKaryotypes}
\title{Make a matrix of karyotypes}
\usage{
makeKaryotypes(
  numCell = 20,
  species = "mouse",
  copies = 2,
  random = FALSE,
  varianceMultiplier = 0.5
)
}
\arguments{
\item{numCell}{The desired number of cells in the matrix.}

\item{species}{The desired species, either mouse or human (20 or 23 chromosome pairs). Otherwise specify the number of chromosome pairs to use.}

\item{copies}{The number of copies for all chromosomes in the karyotype matrix. For random karyotypes this will correspond to the ploidy around which random karyotypes will generated.}

\item{random}{A logical whether to generate random karyotypes (the copies argument will provide the base ploidy (25 percent deviance around the base ploidy will be used).}

\item{varianceMultiplier}{The degree of variance (in standard deviations) in randomly generated karyotypes.}
}
\value{
A matrix of karyotypes according to the input parameters.
}
\description{
This function will create a matrix of karyotypes that is compatible with the CINsim pipeline. Currently supported species are mouse (20 chromosomes) and human (23 chromsomes).
}
\author{
Bjorn Bakker
}
