% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setQmod.R
\name{setQmod}
\alias{setQmod}
\title{Calculate qMod to result in a desired value}
\usage{
setQmod(
  karyotype = NULL,
  desired_value = 0.9,
  chrom_weights = NULL,
  selection_mode = NULL,
  selection_metric = NULL,
  a = NULL,
  b = NULL,
  get_sur_probs = FALSE,
  qMod = 1
)
}
\arguments{
\item{karyotype}{A karyotype provided as a vector (default is 'euploid female mouse')}

\item{desired_value}{The desired value of the parameter based on the provided karyotype.}

\item{chrom_weights}{A vector with chromosome weights to scale importance of one or more chromosome sets.}

\item{selection_mode}{One of three possible modes, "cn_based", "rel_copy", "davoli", for copy number-based, relative copy numbed-based or Davoli-based selection.}

\item{selection_metric}{A fitness metric for the cn_based, rel_copy or davoli selection measure.}

\item{a}{The slope coefficient for the selection paramater.}

\item{b}{The intercept coefficient for the selection paramater.}

\item{get_sur_probs}{A logical whether to return the desired qMod, or return a survival probability.}

\item{qMod}{A custom qMod to modify returned survival probablities.}
}
\value{
A qMod value.
}
\description{
This function will take calculate a qMod value such that a desired value is obtained based
on the selection mode and metric for a specified karyotype (i.e. a euploid karyotype). Alternatively
it can be used to easily calculate survival probabilities for a collection of karyotypes.
}
\author{
Bjorn Bakker
}
